feat: fixed configmap issues
