--- config/samples/02-ref-configs/02-wso2is2.yaml
@@ -1,16 +1,0 @@
-# This pattern assumes that the ConfigMap is created in the same namespace as the WSO2 product deployment.
-
-apiVersion: iam.wso2.com/v1beta1
-kind: Wso2Is
-metadata:
-  name: identity-server-test
-  namespace: wso2-iam-system
-spec:
-  template:
-    annotations:
-      configmapHash: ''
-  replicas: 1
-  configurations:
-    host: identityserver
-    serviceType: ClusterIP
-  tomlConfigFile: wso2is-configmap
