+++ config/samples/archives/standard_multi_node.yaml
