--- config/samples/deployment_mount.yaml
@@ -1,107 +1,0 @@
-apiVersion: iam.wso2.com/v1beta1
-kind: Wso2Is
-metadata:
-  name: identity-server-test
-  namespace: wso2-iam-system
-spec:
-  template:
-    annotations:
-      configmapHash: ''
-  replicas: 2
-  configurations:
-    host: identityserver
-    serviceType: ClusterIP
-  tomlConfig: |
-    [server]  
-    hostname = "identityserver"
-    node_ip = "$env{NODE_IP}"
-    base_path = "https://$ref{server.hostname}"
-
-    [super_admin]
-    username = "admin"
-    password = "admin"
-    create_admin_account = true
-
-    [user_store]
-    type = "database_unique_id"
-
-    [database.identity_db]
-    type = "mysql"
-    url = "jdbc:mysql://mysql:3306/WSO2IS_IDENTITY_DB?autoReconnect=true&amp;allowPublicKeyRetrieval=true&amp;useSSL=false"
-    username = "wso2carbon"
-    password = "wso2carbon"
-    driver = "com.mysql.cj.jdbc.Driver"
-    
-    [database.identity_db.pool_options]
-    validationQuery = "SELECT 1"
-
-    [database.shared_db]
-    type = "mysql"
-    url = "jdbc:mysql://mysql:3306/WSO2IS_SHARED_DB?autoReconnect=true&amp;allowPublicKeyRetrieval=true&amp;useSSL=false"
-    username = "wso2carbon"
-    password = "wso2carbon"
-    driver = "com.mysql.cj.jdbc.Driver"
-    
-    [database.shared_db.pool_options]
-    validationQuery = "SELECT 1"
-
-    [keystore.primary]
-    file_name = "wso2carbon.jks"
-    password = "wso2carbon"
-
-    [truststore]
-    file_name="client-truststore.jks"
-    password="wso2carbon"
-    type="JKS"
-
-    [account_recovery.endpoint.auth]
-    hash= "66cd9688a2ae068244ea01e70f0e230f5623b7fa4cdecb65070a09ec06452262"
-
-    [identity.auth_framework.endpoint]
-    app_password= "dashboard"
-    
-    [transport.https.properties]
-    proxyPort="443"
-    
-    [clustering]
-    membership_scheme = "kubernetes"
-    properties.membershipSchemeClassName = "org.wso2.carbon.membership.scheme.kubernetes.KubernetesMembershipScheme"
-    properties.KUBERNETES_NAMESPACE = "wso2-iam-system"
-    properties.KUBERNETES_SERVICES = "wso2is-service"
-    properties.KUBERNETES_MASTER_SKIP_SSL_VERIFICATION = "true"
-    properties.USE_DNS = "false"
-
-    # The KeyStore which is used for encrypting/decrypting internal data. By default the primary keystore is used as the internal keystore.
-
-    #[keystore.internal]
-    #file_name = "$ref{keystore.primary.file_name}"
-    #type = "$ref{keystore.primary.type}"
-    #password = "$ref{keystore.primary.password}"
-    #alias = "$ref{keystore.primary.alias}"
-    #key_password = "$ref{keystore.primary.key_password}"
-
-    # The KeyStore which is used for tls communication. By default the primary keystore is used as the tls keystore.
-
-    #[keystore.tls]
-    #file_name = "$ref{keystore.primary.file_name}"
-    #type = "$ref{keystore.primary.type}"
-    #password = "$ref{keystore.primary.password}"
-    #alias = "$ref{keystore.primary.alias}"
-    #key_password = "$ref{keystore.primary.key_password}"
-
-    #Google reCAPTCHA settings.
-
-    #[recaptcha]
-    #enabled = true
-    #api_url = "https://www.google.com/recaptcha/api.js"
-    #verify_url = "https://www.google.com/recaptcha/api/siteverify"
-    #site_key = ""
-    #secret_key = ""
-
-    # SMTP email sender settings.
-    #[output_adapter.email]
-    #from_address= "abcd@gmail.com"
-    #username= "abcd"
-    #password= "xxxx"
-    #hostname= "smtp.gmail.com"
-    #port= 587
