+++ controllers/wso2is/pod.go
@@ -1,0 +1,12 @@
+package wso2is
+
+import corev1 "k8s.io/api/core/v1"
+
+// getPodNames returns the pod names of the array of pods passed in
+func getPodNames(pods []corev1.Pod) []string {
+	var podNames []string
+	for _, pod := range pods {
+		podNames = append(podNames, pod.Name)
+	}
+	return podNames
+}
